put landmark code into object form
added object to handle everything landmark related (position, rssi, distance, index, mac address). To that end the object contains an array of structures. Functions were also added to the object to allow for operation on the member variables. 

In addition added plotting functionality such that the plot will contain a line that shows where the robot thinks an observed landmark is when it receives data from the nodes